chore: added command line arguments, removed dead code
