+++ codegen-examples/examples/swebench_agent_run/entry_point.py
@@ -1,5 +1,5 @@
-from codegen.extensions.langchain.utils import SweBenchExample
-from codegen.extensions.swebench.harness import process_one_instance
+from codegen.extensions.swebench.utils import SweBenchExample
+from codegen.extensions.swebench.harness import run_agent_on_entry
 import modal
 
 image = (
@@ -9,13 +9,11 @@ image = (
     .copy_local_dir("../../../", "/root/codegen", ignore=[".venv", "**/.venv", "tests", "**/tests"])
     .run_commands("pip install -e /root/codegen")
 )
-app = modal.App(name="swebench-agent-run", image=image, secrets=[modal.Secret.from_dotenv()])
 
-# Here is an example implementation of setting up an endpoint for receiving webhook events from Linear.
-# The @app.linear.event() decorator takes care of subscribing to the webhook and also unsubscribing when the deployment spun
-# Load environment variables from .env file
+app = modal.App(name="swebench-agent-run", image=image, secrets=[modal.Secret.from_dotenv()])
 
 
 @app.function(timeout=5 * 60)
-async def process_entry(entry: SweBenchExample):
-    return process_one_instance(entry)
+async def run_agent_modal(entry: SweBenchExample):
+    """Modal function to process a single example from the SWE-bench dataset."""
+    return run_agent_on_entry(entry)
