+++ pyproject.toml
@@ -71,7 +71,6 @@ dependencies = [
   "slack-sdk",
   "langchain-anthropic>=0.3.7",
   "lox>=0.12.0",
-  "datasets",
 ]
 
 license = { text = "Apache-2.0" }
