--- src/codegen/extensions/swebench/constants.py
@@ -1,5 +1,0 @@
-from pathlib import Path
-
-PARENT_DIR = Path(__file__).parent
-
-PREDS_DNAME = PARENT_DIR / "predictions"
